Remove references to enableFabricStopAllSurfacesOnTeardown
Summary:
According to our experiments it's not better than "stop surface on unmount" in any way, and might regress some metrics. Unclear why, but if it's not necessary and doesn't seem to help, it doesn't make sense to continue this experiment.

We still have a mechanism on the C++ side to stop outstanding surfaces on teardown that does the same thing.

Changelog: [Internal]

Reviewed By: mdvacca

Differential Revision: D22318864

fbshipit-source-id: 7e678c63e4884382e57d996a7f4c4b7b24c8193a