Ignore exceptions when a Delete instruction is in the same BatchMountItem as its corresponding Create instruction during stopSurface
Summary:
See title. Basically during stopSurface a single BatchMountInstruction can contain both the Create and Delete MountItem for a single view, which will cause *only* the deletion to be executed.

There isn't really a way to prevent this and we're just trying to clean up as aggressively as possible, so we can safely ignore this.

Changelog: [Internal]

Reviewed By: mdvacca

Differential Revision: D22779189

fbshipit-source-id: c44fd736835b04c5de776346ec3d34afa4860199