Create Fabric flavor into RN Tester build system
Summary:
This diff adds a new flavor into the RN Tester gradle system.

The "fabric" flavor is going to be used in the future to:
- determine if Fabric should be enabled or not (at runtime)
- include or exclude fabric C++ code at compile time

I decided to temporarily  reuse the "vm" dimension in order to reduce friction while I'm still iterating on the implementation.
If I created another dimension, then it will force developers to compile RN tester using two flavors:

```
./gradlew  :RNTester:android:app:installHermesFabricDebug
```

I'm planning to refactor this in the near future (T71370706)

changelog: [Internal] Internal

Reviewed By: JoshuaGross

Differential Revision: D23012445

fbshipit-source-id: 1ea6a707ea2bfaca1be567f43b08860407124c17